Sam # The Star Wars Episode where Luke loses his hand.
Checks too see the highest hand, returning as an array of ints the hierarcy.
The straight loops may not work.
Hierarchy is returned in order (TYPE, HIGHCARD, SUIT), unless in case of 2 pair, where  the order is (TYPE, HIGHCARD1, HIGHCARD2,SUIT)
The Hand class gets the "hand" of a player--aka the best of the 21 possibilities. Then, a player's hand can be compared.