Sam # The number of us at code day
Checks too see the highest hand, returning as an array of ints the hierarcy.
The straight loops may not work.
Hierarchy is returned in order (TYPE, MATCH NUMBER, HIGHCARD), unless in case of 2 pair, where  the order is (TYPE, MATCH1, MATCH2,HIGHCARD)
The Hand class gets the "hand" of a player--aka the best of the 21 possibilities. Then, a player's hand can be compared. It comes quite in handy;)
The hand class now functions for determining the best hand very rapidly-- and I have to HAND it to my pals Alekz and Nicolo.