Sam # to dislike intensely.substring(1,3);
Checks too see the highest hand, returning as an array of ints the hierarcy.
The straight loops may not work.
Hierarchy is returned in order (TYPE, MATCH NUMBER, HIGHCARD), unless in case of 2 pair, where  the order is (TYPE, MATCH1, MATCH2,HIGHCARD)
The Hand class gets the "hand" of a player--aka the best of the 21 possibilities. Then, a player's hand can be compared. It comes quite in handy;)
The hand class now functions for determining the best hand very rapidly-- and I have to HAND it to my pals Alekz and Nicolo.
PokerGame now has a way of checking the winning hand, and a way to change that array into a string. MAKE SURE TO PUT THE CHECKING FOR WINNER AS THE ARGUMENTS OF TURNING INTO A STRING!!
Sam will not be available for the next hour or so. If you would like to reach him, just shout "SAM!!!!!" really loudly.